perf(main.py): inline file writing and error handling
Adjusted script to write files as they are read so with error handling in case a file bugs. This will make it so the program doesnt halt if an error occurs, but rather the file is skipped.

An error occured where a file under 1 second had no audible audio, the model inserted korean symbls however the ecoding was improper and the programming halted.