fix(main.py): remove schedule
Removed the time schedule for running. It is currently not working properly and a solution will be put in place at a later time.

BREAKING CHANGE: The program will need to be manually turned on when it is intended to run, the alternative is to leave it running and every 5 minutes the program will go to sleep if there is no job to be done.